[GEOS-9826] Remove un-necessary casts from code, add a QA rule to enforce it

+++ src/extension/netcdf-out/src/main/java/org/geoserver/wcs/responses/AbstractNetCDFEncoder.java
@@ -618,7 +618,7 @@ public abstract class AbstractNetCDFEncoder implements NetCDFEncoder {
                     sample = (int) unitConverter.convert(sample);
                 }
                 if (dataPacker != null) {
-                    sample = dataPacker.pack((double) sample);
+                    sample = dataPacker.pack(sample);
                 }
                 setIntegerSample(netCDFDataType, matrix, matrixIndex, sample);
                 break;
